main.cpp をファイル分割（リファクタリングのみ）
- config.h: 定数・ピン・ゲイン・タイミング設定
- shared_state.h: CtrlCommand/TelemetryData 型定義・キュー extern 宣言
- control_task.cpp: 制御ループ（PID + Mahony + DXL 駆動）
- ui_task.cpp: タッチ UI + テレメトリ Serial 出力
- main.cpp: setup() + タスク生成のみ ~75 行に縮小
- 動作変更なし
